test_requirements.txt
django-nose
coverage
django-coverage
django-libs
ipdb
flake8
fabric
factory_boy
mock
